+++ packages/houdini-react/src/runtime/routing/Router.tsx
@@ -15,6 +15,8 @@ import { DocumentHandle, useDocumentHandle } from '../hooks/useDocumentHandle'
 import { useDocumentStore } from '../hooks/useDocumentStore'
 import { SuspenseCache, suspense_cache } from './cache'
 
+type PageComponent = React.ComponentType<{ url: string }>
+
 const PreloadWhich = {
 	component: 'component',
 	data: 'data',
@@ -467,7 +469,7 @@ export function RouterContextProvider({
 	client: HoudiniClient
 	cache: Cache
 	artifact_cache: SuspenseCache<QueryArtifact>
-	component_cache: SuspenseCache<(props: any) => React.ReactElement>
+	component_cache: SuspenseCache<PageComponent>
 	data_cache: SuspenseCache<DocumentStore<GraphQLObject, GraphQLVariables>>
 	ssr_signals: PendingCache
 	last_variables: LRUCache<GraphQLVariables>
@@ -522,7 +524,7 @@ type RouterContext = {
 
 	// We also need a cache for component references so we can avoid suspending
 	// when we load the same page multiple times
-	component_cache: SuspenseCache<(props: any) => React.ReactElement>
+	component_cache: SuspenseCache<PageComponent>
 
 	// Pages need a way to wait for data
 	data_cache: SuspenseCache<DocumentStore<GraphQLObject, GraphQLVariables>>
@@ -791,7 +793,7 @@ function usePreload({ preload }: { preload: (url: string, which: PreloadWhichVal
 
 export type RouterCache = {
 	artifact_cache: SuspenseCache<QueryArtifact>
-	component_cache: SuspenseCache<(props: any) => React.ReactElement>
+	component_cache: SuspenseCache<PageComponent>
 	data_cache: SuspenseCache<DocumentStore<GraphQLObject, GraphQLVariables>>
 	last_variables: LRUCache<GraphQLVariables>
 	ssr_signals: PendingCache
@@ -807,7 +809,7 @@ export function router_cache({
 }: {
 	pending_queries?: string[]
 	artifacts?: Record<string, QueryArtifact>
-	components?: Record<string, (props: any) => React.ReactElement>
+	components?: Record<string, PageComponent>
 	initialData?: Record<string, DocumentStore<GraphQLObject, GraphQLVariables>>
 	initialVariables?: Record<string, GraphQLVariables>
 	initialArtifacts?: Record<string, QueryArtifact>