+++ packages/houdini-svelte/src/plugin/transforms/kit/session.test.ts
@@ -42,11 +42,9 @@ test('export const load', async function () {
 
 		export const load = loadFlash(async event => {
 		    "some random stuff that's valid javascript";
-		    const __houdini__vite__plugin__return__value__ = {};
-
 		    return {
 		        ...buildSessionObject(event),
-		        ...__houdini__vite__plugin__return__value__
+		        ...{}
 		    };
 		});
 	`)
@@ -76,11 +74,9 @@ test('adds load to +layout.server.js', async function () {
 		import { buildSessionObject } from "$houdini/plugins/houdini-svelte/runtime/session";
 
 		export async function load(event) {
-		    const __houdini__vite__plugin__return__value__ = {};
-
 		    return {
 		        ...buildSessionObject(event),
-		        ...__houdini__vite__plugin__return__value__
+		        ...{}
 		    };
 		}
 	`)
@@ -105,13 +101,12 @@ test('modifies existing load +layout.server.js', async function () {
 
 		export async function load(event) {
 		    "some random stuff that's valid javascript";
-		    const __houdini__vite__plugin__return__value__ = {
-		        hello: "world"
-		    };
-
 		    return {
 		        ...buildSessionObject(event),
-		        ...__houdini__vite__plugin__return__value__
+
+		        ...{
+		            hello: "world"
+		        }
 		    };
 		}
 	`)
@@ -132,11 +127,9 @@ test('modifies existing load +layout.server.js - no return', async function () {
 
 		export async function load(event) {
 		    "some random stuff that's valid javascript";
-		    const __houdini__vite__plugin__return__value__ = {};
-
 		    return {
 		        ...buildSessionObject(event),
-		        ...__houdini__vite__plugin__return__value__
+		        ...{}
 		    };
 		}
 	`)
@@ -157,13 +150,12 @@ test('modifies existing load +layout.server.js - satisfies operator', async func
 		import type { LayoutServerLoad } from "./$types";
 
 		export const load = (event => {
-		    const __houdini__vite__plugin__return__value__ = ({
-		        test: "Hello"
-		    });
-
 		    return {
 		        ...buildSessionObject(event),
-		        ...__houdini__vite__plugin__return__value__
+
+		        ...({
+		            test: "Hello"
+		        })
 		    };
 		}) satisfies LayoutServerLoad;
 	`)
@@ -194,13 +186,12 @@ test('modifies existing load +layout.server.js - rest params', async function ()
 
 		    console.log(foo);
 
-		    const __houdini__vite__plugin__return__value__ = {
-		        some: "value"
-		    };
-
 		    return {
 		        ...buildSessionObject(event),
-		        ...__houdini__vite__plugin__return__value__
+
+		        ...{
+		            some: "value"
+		        }
 		    };
 		}
 	`)
@@ -231,13 +222,12 @@ test('modifies existing load +layout.server.js - const arrow function', async fu
 
 		    console.log(foo);
 
-		    const __houdini__vite__plugin__return__value__ = {
-		        some: "value"
-		    };
-
 		    return {
 		        ...buildSessionObject(event),
-		        ...__houdini__vite__plugin__return__value__
+
+		        ...{
+		            some: "value"
+		        }
 		    };
 		};
 	`)
@@ -268,13 +258,12 @@ test('modifies existing load +layout.server.js - const function', async function
 
 		    console.log(foo);
 
-		    const __houdini__vite__plugin__return__value__ = {
-		        some: "value"
-		    };
-
 		    return {
 		        ...buildSessionObject(event),
-		        ...__houdini__vite__plugin__return__value__
+
+		        ...{
+		            some: "value"
+		        }
 		    };
 		};
 	`)
@@ -292,13 +281,12 @@ test('modifies existing load +layout.server.js - implicit return', async functio
 		import { buildSessionObject } from "$houdini/plugins/houdini-svelte/runtime/session";
 
 		export const load = event => {
-		    const __houdini__vite__plugin__return__value__ = ({
-		        hello: "world"
-		    });
-
 		    return {
 		        ...buildSessionObject(event),
-		        ...__houdini__vite__plugin__return__value__
+
+		        ...({
+		            hello: "world"
+		        })
 		    };
 		};
 	`)
@@ -309,12 +297,92 @@ test('passes session from root client-side layout', async function () {
 
 	expect(result).toMatchInlineSnapshot(`
 		export async function load(event) {
-		    const __houdini__vite__plugin__return__value__ = {};
+		    return {
+		        ...event.data,
+		        ...{}
+		    };
+		}
+	`)
+})
+
+test('augments load function in root layout load', async function () {
+	const result = await test_transform_js(
+		'src/routes/+layout.js',
+		`
+		export const load = async ({ url }) => {
+			console.log('routes/+layout.js start');
+			if (!browser) return;
+
+			console.log('this should only run in the browser');
+		};
+	`
+	)
+
+	expect(result).toMatchInlineSnapshot(`
+		export const load = async event => {
+		    let {
+		        url
+		    } = event;
+
+		    console.log("routes/+layout.js start");
+
+		    if (!browser) return {
+		        ...event.data,
+		        ...{}
+		    };
+
+		    console.log("this should only run in the browser");
 
 		    return {
 		        ...event.data,
-		        ...__houdini__vite__plugin__return__value__
+		        ...{}
 		    };
-		}
+		};
+	`)
+})
+
+test('call expression assignment for load function', async function () {
+	const result = await test_transform_js(
+		'src/routes/+layout.js',
+		`
+
+		export const load = someFn()
+	`
+	)
+
+	expect(result).toMatchInlineSnapshot(`
+		export const load = event => {
+		    const result = houdini__intermediate__load__(event);
+
+		    return {
+		        ...event.data,
+		        ...result
+		    };
+		};
+
+		const houdini__intermediate__load__ = someFn();
+	`)
+})
+
+test('value assignment for load function', async function () {
+	const result = await test_transform_js(
+		'src/routes/+layout.js',
+		`
+
+		export const load = someFn
+	`
+	)
+
+	expect(result).toMatchInlineSnapshot(`
+		export const load = event => {
+		    const result = houdini__intermediate__load__(event);
+
+		    return {
+		        ...event.data,
+		        ...result
+		    };
+		};
+
+		const houdini__intermediate__load__ = someFn;
 	`)
 })