+++ packages/houdini/src/codegen/generators/artifacts/tests/artifacts.test.ts
@@ -641,7 +641,7 @@ test('paginate over unions', async function () {
 		export default {
 		    "name": "TestQuery",
 		    "kind": "HoudiniQuery",
-		    "hash": "1d660603f2610ac59b38fd74f0b429f0c9c2cec3c64459cb222e7e26c7905e66",
+		    "hash": "c288a4c72d1d7ce2786cc5a4dcbb4263eae036fe4142f127d0b0cfab9895c45f",
 
 		    "refetch": {
 		        "path": ["entitiesByCursor"],
@@ -668,6 +668,7 @@ test('paginate over unions', async function () {
 		    edges {
 		      cursor
 		      node {
+		        __typename
 		        __typename
 		      }
 		    }
@@ -4749,7 +4750,7 @@ describe('mutation artifacts', function () {
 			export default {
 			    "name": "TestQuery",
 			    "kind": "HoudiniQuery",
-			    "hash": "6fe0aeaa708161553cd04645834b38c4ce625fce10c46056efcff9a97988d358",
+			    "hash": "6ca70e636d72875139c7d7bb34b4a9f59d88e2b518613aa136fda4b0ed539936",
 
 			    "refetch": {
 			        "path": ["usersByCursor"],
@@ -4774,6 +4775,7 @@ describe('mutation artifacts', function () {
 			      cursor
 			      node {
 			        __typename
+			        id
 			      }
 			    }
 			    pageInfo {
@@ -5001,7 +5003,7 @@ describe('mutation artifacts', function () {
 			export default {
 			    "name": "TestQuery",
 			    "kind": "HoudiniQuery",
-			    "hash": "6fe0aeaa708161553cd04645834b38c4ce625fce10c46056efcff9a97988d358",
+			    "hash": "6ca70e636d72875139c7d7bb34b4a9f59d88e2b518613aa136fda4b0ed539936",
 
 			    "refetch": {
 			        "path": ["usersByCursor"],
@@ -5026,6 +5028,7 @@ describe('mutation artifacts', function () {
 			      cursor
 			      node {
 			        __typename
+			        id
 			      }
 			    }
 			    pageInfo {