+++ packages/houdini/src/codegen/generators/artifacts/tests/pagination.test.ts
@@ -30,7 +30,7 @@ test('pagination arguments stripped from key', async function () {
 		export default {
 		    "name": "PaginatedFragment",
 		    "kind": "HoudiniFragment",
-		    "hash": "fd66badc676c41d3a2449afc85e1901f9fc1580e030c7658acff1ce4fb1b4d3f",
+		    "hash": "d06fb44a3b945dac99c82f58b6d422248493cc83dbdef0e159de034dc541b6f9",
 
 		    "refetch": {
 		        "path": ["friendsByCursor"],
@@ -60,6 +60,7 @@ test('pagination arguments stripped from key', async function () {
 		      cursor
 		      node {
 		        __typename
+		        id
 		      }
 		    }
 		    pageInfo {
@@ -238,7 +239,7 @@ test('pagination arguments stays in key as it s a SinglePage Mode', async functi
 		export default {
 		    "name": "PaginatedFragment",
 		    "kind": "HoudiniFragment",
-		    "hash": "fd66badc676c41d3a2449afc85e1901f9fc1580e030c7658acff1ce4fb1b4d3f",
+		    "hash": "d06fb44a3b945dac99c82f58b6d422248493cc83dbdef0e159de034dc541b6f9",
 
 		    "refetch": {
 		        "path": ["friendsByCursor"],
@@ -268,6 +269,7 @@ test('pagination arguments stays in key as it s a SinglePage Mode', async functi
 		      cursor
 		      node {
 		        __typename
+		        id
 		      }
 		    }
 		    pageInfo {
@@ -567,7 +569,7 @@ test('cursor as scalar gets the right pagination query argument types', async fu
 		export default {
 		    "name": "ScalarPagination",
 		    "kind": "HoudiniQuery",
-		    "hash": "52d1832ab8d47b638e7f73bf90390ffd2e3a41ec194f0cc821ebe51ad792d771",
+		    "hash": "476b5ddafe8444d071fb07db5852f9f364ff8ef9cf5ba361f0d0685c4e9bd241",
 
 		    "refetch": {
 		        "path": ["user", "friendsByCursorScalar"],
@@ -605,6 +607,7 @@ test('cursor as scalar gets the right pagination query argument types', async fu
 		        cursor
 		        node {
 		          __typename
+		          id
 		        }
 		      }
 		      pageInfo {
@@ -855,7 +858,7 @@ test("sibling aliases don't get marked", async function () {
 		export default {
 		    "name": "PaginatedFragment",
 		    "kind": "HoudiniFragment",
-		    "hash": "5efa295fd82f7f78f41f944f345fbea76fbf5c5988ecfed7e0ac3ae757423f5b",
+		    "hash": "7665f8fd8a0af62317c95868d5e018001152de84c95f653f5caba6b515e367d6",
 
 		    "refetch": {
 		        "path": ["friendsByCursor"],
@@ -892,6 +895,7 @@ test("sibling aliases don't get marked", async function () {
 		      cursor
 		      node {
 		        __typename
+		        id
 		      }
 		    }
 		    pageInfo {